// src/util/enums.ts
export enum UserRole {
  ADMIN = 'Admin',
  USER = 'User',
}
export enum Language {
  EN = 'en',
  ES = 'es',
  FR = 'fr',
}

export enum JokeSource {
  API = 'api',
  USER = 'user',
  FALLBACK = 'fallback',
}

export enum JokeStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  REJECTED = 'rejected',
  FLAGGED = 'flagged',
}